refactor: move to flat test files inside test dir
